fiddeled with the initialiser code

+++ cmd/reliably/init/cmd.go
@@ -2,6 +2,7 @@ package init
 
 import (
 	"bufio"
+	"fmt"
 	"os"
 	"strings"
 
@@ -15,6 +16,7 @@ import (
 var (
 	manifestPath        string
 	supportedExtensions = []string{".yaml", ".json"}
+	autoInitialise      bool
 )
 
 func NewCommand() *cobra.Command {
@@ -26,16 +28,26 @@ func NewCommand() *cobra.Command {
 	}
 
 	cmd.Flags().StringVarP(&manifestPath, "manifest-file", "f", "reliably.yaml", "the location of the manifest file")
+	cmd.Flags().BoolVarP(&autoInitialise, "auto-init", "i", false, "Auto-initialise the manifest")
 
 	return &cmd
 }
 
-func run(_ *cobra.Command, _ []string) {
+func run(_ *cobra.Command, args []string) {
 	validateFilePath()
+	scanner := bufio.NewScanner(os.Stdin)
 
 	m := manifest.New()
 
-	populateManifest(m)
+	if !autoInitialise {
+		if _, err := os.Stat(manifestPath); err == nil {
+			if !question.WithBoolAnswer(scanner, fmt.Sprintf("File '%s' already exists. Do you want to replace it (y/n)?", manifestPath)) {
+				return
+			}
+		}
+
+		populateManifestInteractively(m, scanner)
+	}
 
 	f, err := os.Create(manifestPath)
 	if err != nil {
@@ -58,20 +70,18 @@ func validateFilePath() {
 	log.Fatalf("manifest file must have one of the these extensions: %v", supportedExtensions)
 }
 
-func populateManifest(m *manifest.Manifest) {
-	scanner := bufio.NewScanner(os.Stdin)
-
+func populateManifestInteractively(m *manifest.Manifest, scanner *bufio.Scanner) {
 	m.App.Name = question.WithStringAnswer(scanner, "What is the name of your application?")
 	m.App.Owner = question.WithStringAnswer(scanner, "Who owns this app? Its a good idea to put an email address in here!")
 	m.App.Repository = question.WithStringAnswer(scanner, "What is the URL to the repository for this app?")
 
-	if question.WithBoolAnswer(scanner, "Are you using Continuous Integration?") {
+	if question.WithBoolAnswer(scanner, "Are you using Continuous Integration? (y/n)") {
 		m.CI.Type = question.WithStringAnswer(scanner, "What type of CI are you using?")
 	} else {
 		m.CI = nil
 	}
 
-	if question.WithBoolAnswer(scanner, "Are you building something that will be provided to customers 'as a service'?") {
+	if question.WithBoolAnswer(scanner, "Are you building something that will be provided to customers 'as a service'? (y/n)") {
 		m.ServiceLevel.Availability = question.WithFloat64Answer(scanner, "What percentage of availability do you want your application to have?", 0, 100)
 		m.ServiceLevel.ErrorBudgetPercent = question.WithFloat64Answer(scanner, "What percentage of requests to your service is it ok to have fail? This will be your 'error budget'.", 0, 100)
 		m.ServiceLevel.Latency = question.WithDurationAnswer(scanner, "What is the maximum request-response latency you want from this service")
@@ -79,7 +89,7 @@ func populateManifest(m *manifest.Manifest) {
 		m.ServiceLevel = nil
 	}
 
-	if question.WithBoolAnswer(scanner, "Will your app be hosted on a commercial platform?") {
+	if question.WithBoolAnswer(scanner, "Will your app be hosted on a commercial platform? (y/n)") {
 		m.Hosting = &manifest.Hosting{
 			Provider: question.WithStringAnswer(scanner, "What is the name of the provider?"),
 		}
@@ -87,7 +97,7 @@ func populateManifest(m *manifest.Manifest) {
 		m.Hosting = nil
 	}
 
-	if question.WithBoolAnswer(scanner, "Does your application have 'service level' dependencies?") {
+	if question.WithBoolAnswer(scanner, "Does your application have 'service level' dependencies? (y/n)") {
 		deps := make([]*manifest.Dependency, 0)
 
 		addMore := true
@@ -98,7 +108,7 @@ func populateManifest(m *manifest.Manifest) {
 
 			deps = append(deps, d)
 
-			addMore = question.WithBoolAnswer(scanner, "Do you want to add another dependency?")
+			addMore = question.WithBoolAnswer(scanner, "Do you want to add another dependency? (y/n)")
 		}
 
 		m.Dependencies = deps
@@ -106,7 +116,7 @@ func populateManifest(m *manifest.Manifest) {
 		m.Dependencies = nil
 	}
 
-	if question.WithBoolAnswer(scanner, "Are you using Infrastucture as Code?") {
+	if question.WithBoolAnswer(scanner, "Are you using Infrastucture as Code? (y/n)") {
 		m.IAC = &manifest.IAC{
 			Type: question.WithStringAnswer(scanner, "What IAC provider are you using (terraform, ARM templates, CDK, etc...)?"),
 			Root: question.WithStringAnswer(scanner, "Where is the root folder for your IAC code?"),
